Styles updated for all the screens, color scheem updated for all the screens, Json and Style sheets sepatared in their respective folders, style bugs fixed: Crousel images updated:Moeez Ahmed
