+++ Styles/AddAddressStyles.js
@@ -1,0 +1,86 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: 'white',
+        borderTopLeftRadius: 20,
+        borderTopRightRadius: 20,
+    },
+    scrollContainer: {
+        flexGrow: 1,
+        paddingBottom: 20,
+        justifyContent: 'space-between'
+    },
+    headerContainer: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        padding: 15,
+        borderBottomWidth: 1,
+        borderBottomColor: '#f0f0f0'
+    },
+    closeButton: {
+        padding: 10,
+    },
+    headerTitle: {
+        fontSize: 18,
+        fontWeight: 'bold',
+        marginLeft: 10,
+        color: '#FF2B85'
+    },
+    formContainer: {
+        padding: 15
+    },
+    label: {
+        marginTop: 15,
+        marginBottom: 5,
+        fontWeight: 'bold'
+    },
+    input: {
+        borderWidth: 1,
+        borderColor: '#e0e0e0',
+        borderRadius: 8,
+        padding: 10,
+        marginBottom: 10
+    },
+    typeContainer: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        marginBottom: 15
+    },
+    typeButton: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        borderWidth: 1,
+        borderColor: '#e0e0e0',
+        borderRadius: 8,
+        padding: 10,
+        width: '30%'
+    },
+    selectedTypeButton: {
+        backgroundColor: '#FF2B85',
+        borderColor: '#e52850'
+    },
+    typeButtonText: {
+        marginLeft: 5
+    },
+    selectedTypeButtonText: {
+        color: 'white'
+    },
+    textArea: {
+        height: 100,
+        textAlignVertical: 'top'
+    },
+    saveButton: {
+        backgroundColor: '#FF2B85',
+        padding: 15,
+        borderRadius: 8,
+        alignItems: 'center',
+        marginTop: 20
+    },
+    saveButtonText: {
+        color: 'white',
+        fontWeight: 'bold',
+        fontSize: 16
+    }
+});
