+++ Styles/AddOnPageStyles.js
@@ -3,7 +3,7 @@ import { StyleSheet } from "react-native";
 const styles = StyleSheet.create({
     container: {
         flex: 1,
-        backgroundColor: "#f5f5f5",
+        backgroundColor: "#FF2B85",
         paddingTop: 40,
     },
     header: {
