+++ Styles/AvatarStyles.js
@@ -1,0 +1,36 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  avatar: {
+    borderRadius: 100,
+    overflow: 'hidden',
+    maxWidth: '100%',
+  },
+  image: {
+    objectFit: 'cover',
+    paddingTop: 0,
+  },
+  noImage: {
+    backgroundColor: '#333',
+    borderWidth: 1,
+    borderStyle: 'solid',
+    borderColor: 'rgb(200, 200, 200)',
+    borderRadius: 5,
+  },
+  button: {
+    marginTop: 10,
+    padding: 10,
+    backgroundColor: '#4630EB',
+    borderRadius: 5,
+  },
+  buttonText: {
+    color: 'white',
+    fontWeight: 'bold',
+  },
+})
+
+export default styles
